feat: implement opts.library.blacklist (#148)
This PR implements #148 by adding a new option (`opts.library.blacklist`) which is a string[] that consists
of plugin names (just as `opts.library.plugins`) which will not be included in the final path.

Note: This overrides `opts.library.plugins` in that, if a plugin is included in the blacklist, it *will not*
be included, even if included in the `plugins` list.