fix config.root() so it returns a string or nil
config.root() should return either string or nil, but
vim.loop.fs_realpath returns a 3-tuple (see `:help vim.uv.fs_realpath`
and `:help luv-error-handling`). By adding parentheses around the
function call, we force root() to return a string or nil, and also
types() to return string or nil.

This fixes an error when trying to use rocks.nvim [1] to install
neodev.nvim. Since the `types/` folder is nowhere to be found when
installing neodev.nvim with luarocks, the call to fs_realpath returns an
error, and stops the execution after any call to config.types().

[1] https://github.com/nvim-neorocks/rocks.nvim